qdio: prevent kernel bug message in interrupt handler
Patchwork-id: 23331
O-Subject: [RHEL6 PATCH 4/5] [s390] qdio: prevent kernel bug message in
	interrupt handler
Bugzilla: 565542

Description
-----------
Setting a qdio chpid offline during a qdio device shutdown
can lead to a failing ccw_device_halt which leaves the qdio
device in the stopped state.  If an interrupt occurs before
the final cleanup is done the unhandled stopped state triggers
a kernel bug message.

Add the stopped state to the known states in the interrupt handler
and prevent the kernel bug message for the stopped state.

Bugzilla
--------
BZ 565542
https://bugzilla.redhat.com/show_bug.cgi?id=565542

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=959153d34544b7237bad263e73a5abdf94fc7722

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>