virtio: console: Error out if we can't allocate buffers for control queue
Patchwork-id: 23342
O-Subject: [PATCH 08/10] virtio: console: Error out if we can't allocate buffers
	for control queue
Bugzilla: 566391
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>

With MULTIPORT support, the control queue is an integral part of the
functioning of the device. If we can't get any buffers allocated, the
host won't be able to relay important information and the device may not
function as intended.

Ensure 'probe' doesn't succeed until the control queue has at least one
buffer allocated for its ivq.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>