megaraid_sas: fix for 32bit apps
Patchwork-id: 23394
O-Subject: [RHEL6 PATCH] bz#559941 megaraid_sas: fix for 32bit apps
Bugzilla: 559941
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz#559941

This should resolve a problem introduced by the last
driver update to v4.17.
Part of this patch in function megasas_mgmt_fw_ioctl
should have resolved a problem with a 64bit pointer
stored in a u32.
-      u32 *sense_ptr;
+      unsigned long *sense_ptr;

...
          sense_ptr =
-            (u32 *) ((unsigned long)cmd->frame + ioc->sense_off);
+            (unsigned long *) ((unsigned long)cmd->frame + ioc->sense_off);
          *sense_ptr = sense_handle;
...
-      sense_ptr = (u32 *) ((unsigned long)ioc->frame.raw +
-                             ioc->sense_off);
+      sense_ptr = (unsigned long *) ((unsigned long)ioc->frame.raw +
+                       ioc->sense_off);

       if (copy_to_user((void __user *)((unsigned long)(*sense_ptr)),
                        sense, ioc->sense_len)) {
...
This on the other side caused problems for 32bit
programs in 64 bit os, for example Dell's 'open manage'(omsa).

The patch for this is now upstrem -
[SCSI] megaraid_sas: fix for 32bit apps b3dc1a212e5167984616445990c76056034f8eeb
a similar patch is in rhel5.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>