add suport for rhts testing
Patchwork-id: 23453
O-Subject: [RFC][RHEL-6 PATCH] internal git:  add suport for rhts testing
Bugzilla:

This is the core infrastructure piece I have been working on to make it
easier for developers to run automated tests through RHTS.

It adds a couple of new commands namely, rh-brew-test and rh-test.
The first test does the same thing as rh-brew except it takes the task
number of your brew job and passes it into RHTS to use as the kernel to test
with.

The second test, rh-test only does the testing part and thus requires the
task number to be passed in with TASK=1234.

The results of these tests should be emailed to you.

Both of these commands need to be told what test to use.  For this I mimic'd
what was used in the versioning scheme (ie localversion) and used a file
called 'tests' (ie <git-tree>/tests).  The supported tests to choose from
can be found in the redhat/tests/ directory (don't use the .xml in the
'tests' file).

So far I really don't have much in the way of tests.  I have something for
filesystem folks and working on putting something together for the
networking folks.

The idea behind these tests is to have something various groups can use to
validate their patches have not broken core functionality within their
subsystem.  Most developers don't have a comprehensive testsuite and as a
result their testing isn't as thorough as it should be.  This is an attempt
to create a more robust and easy to use test harness.  I am hoping that if
you are willing to wait for brew to finish, then waiting an extra hour for
testing won't be that painful.

Unfortunately, I need more tests for this to be useful.  If anyone has a
group of tests they would like to run but do not know how to integrate it
into RHTS or understand how flexibly RHTS can be to accomodate almost any
automated tests scenario, please contact me and I will work with you.

Also note the kernel QE team has offered their help in writing tests and can
probably be a huge help in growing the tests but they need to get a feel of
the developers needs.  So again, if you have ideas and some basic tests,
please contact me.

The patch below adds a Makefile.test file to do all the magic.  Some tweaks
were made in redhat/Makefile to not only include Makefile.test, but to
capture the brew output to find the task number to pass into RHTS.

Note: this should not affect any current commands

Please review.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>