qeth: avoid recovery during device online setting
Patchwork-id: 23456
O-Subject: [RHEL6 PATCH 1/1] [s390] qeth: avoid recovery during device online
	setting
Bugzilla: 568781
RH-Acked-by: David S. Miller <davem@redhat.com>

Description
-----------
A kernel panic may occur while qeth recovery is running.

If a qeth device is set online, several initialisation steps
are performed. If a failure in one of these steps occurs,
the qeth device is reset into DOWN state. If due to the
failure a qeth recovery is scheduled and started in another
thread, this might cause all kinds of conflicts, even a
kernel panic.

Forbid scheduling of a qeth recovery while online processing
is performed till the card is in state SOFTSETUP.

Bugzilla
--------
BZ 568781
https://bugzilla.redhat.com/show_bug.cgi?id=568781

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.34.
It is already queued in net-next:
http://git.kernel.org/?p=linux/kernel/git/davem/net-next-2.6.git;a=commitdiff;h=84b66683dba002f4cce2aaf78bf45debfaf22795

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.