disable all the filesystems we aren't going to support and fix ext3/4
Patchwork-id:
O-Subject: [kernel team] [RHEL6 PATCH] disable all the filesystems we aren't going to support and fix ext3/4
Bugzilla: 568789

Hello,

This patch disables all of the filesystems we didn't support in RHEL5 for RHEL6,
and it makes ext3 and 4 be built as modules instead of into the kernel.  This is
in reference to BZ#568789.  Thanks,

Josef

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>