Change RPM name format for custom builds
Patchwork-id: 23472
O-Subject: [RHEL 6] Change RPM name format for custom builds
Bugzilla:
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Anton Arapov <Anton@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>

I ran into an issue where I created a 'localversion' file with the
contents 'bz560077_v1'.  When doing a 'make rh-brew' the SRPM is created
with the filename of kernel-2.6.32-17.bz560077_v1.el6.src.rpm.

The problem occurs after creating a kernel and then
going to install that kernel.  yum think that the created kernel is an
older kernel than the already existing 'kernel-2.6.32-17.el6' RPM, since
'bz560077_v1' is earlier in the alphabet than 'el6'.

This patch changes the filename so that the localversion buildid will be
placed at the end of the RPM name.  So now my RPM will have the name
'kernel-2.6.32-17.el6.bz560077_v2' which yum thinks is newer than the
default installed kernel.  This is also the same as how it currently
works in RHEL 5.

I tested my changes and was able to succesfully do an rh-brew and create
a SRPM and build in brew successfully the RPMs.

To recap:

Old name format:
kernel-2.6.32-XX.LOCALVERSION.el6

New name format:
kernel-2.6.32-XX.el6.LOCALVERSION

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>