scsi_dh_emc: fix mode select setup
Patchwork-id: 23496
O-Subject: [PATCH RHEL 6.0]: scsi_dh_emc: fix mode select setup
Bugzilla: 570685
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for BZ: 570685.

The patch fixes:
- setting of the length for mode selects.
- setting of the data direction for mode select 10.

This patch is upstream here:
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=6c71dcb28ff9b63b814a0b76a256f5dae08d3e0d

EMC had tested the patch when we added this for RHEL 5 and when
I sent upstream. It is not already in RHEL6 because there was a
mixup upstream and it is just now getting merged.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>