revert cc2f92ad1d0e03fe527e8ccfc1f918c368964dc8
Patchwork-id: 23508
O-Subject: [RHEL6 PATCH] kernel/time: revert
	cc2f92ad1d0e03fe527e8ccfc1f918c368964dc8
Bugzilla: 567551
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Zachary Amsden <zamsden@redhat.com>

BZ#567551

Short version: this patch reverts cc2f92ad1d0e03fe527e8ccfc1f918c368964dc8
based on a revert also done upstream matching exactly the problem we're seeing
in BZ#567551 and I wasn't able to test so far.

Long version:
This patch needs more explanation than the usual. The BZ in question reports a
problem during installation in which cp fails to preserve the modification
timestamp while copying keymap files in order to build the initrd. After
further analysis, it was found that the source file indeed had 1s worth in
nanoseconds in atime timespec. It's hard to reproduce and I only could
reproduce it during installation. Now here's the catch.

cc2f92ad1d0e03fe527e8ccfc1f918c368964dc8 is the backport of upstream's
7bc7d637452383d56ba4368d4336b0dde1bb476d, which removes xtime_cache. In both
upstream and in the backport, it was done as cleanup since xtime_cache wasn't
useful anymore. On the upstream's description of the revert:

    This reverts commit 7bc7d637452383d56ba4368d4336b0dde1bb476d, as
    requested by John Stultz. Quoting John:

     "Petr TitÄ<9B>ra reported an issue where he saw odd atime regressions with
      2.6.33 where there were a full second worth of nanoseconds in the
      nanoseconds field.

      He also reviewed the time code and narrowed down the problem: unhandled
      overflow of the nanosecond field caused by rounding up the
      sub-nanosecond accumulated time.

      Details:

       * At the end of update_wall_time(), we currently round up the
      sub-nanosecond portion of accumulated time when storing it into xtime.
      This was added to avoid time inconsistencies caused when the
      sub-nanosecond portion was truncated when storing into xtime.
      Unfortunately we don't handle the possible second overflow caused by
      that rounding.

       * Previously the xtime_cache code hid this overflow by normalizing the
      xtime value when storing into the xtime_cache.

       * We could try to handle the second overflow after the rounding up, but
      since this affects the timekeeping's internal state, this would further
      complicate the next accumulation cycle, causing small errors in ntp
      steering. As much as I'd like to get rid of it, the xtime_cache code is
      known to work.

       * The correct fix is really to include the sub-nanosecond portion in the
      timekeeping accessor function, so we don't need to round up at during
      accumulation. This would greatly simplify the accumulation code.
      Unfortunately, we can't do this safely until the last three
      non-GENERIC_TIME arches (sparc32, arm, cris) are converted  (those
      patches are in -mm) and we kill off the spots where arches set xtime
      directly. This is all 2.6.34 material, so I think reverting the
      xtime_cache change is the best approach for now.

      Many thanks to Petr for both reporting and finding the issue!"

which matches exaclty what we're seeing here. The problem is that to test this
patch I need a new compose (it must be the kernel that anaconda is using during
installation, not later). Ticket #60254 was filled with eng-ops to get this
done. Further tries to get this running other way around didn't work so far.

Eric Sandeen checked for possible problems on ext4 handling of nanosecond
timestamps and couldn't find any.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>