update cifs client code to latest upstream code
Patchwork-id: 23525
O-Subject: [RHEL6.0 PATCH] BZ#562788: cifs: update cifs client code to latest
	upstream code
Bugzilla: 562788
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

This patchset is a rollup of all of the latest changesets that have gone
upstream. There are a number of smallish bugfixes and a couple of larger
sets -- one to fix a problem in the extended attribute parsing code, and
one to fix some problems in the cache revalidation code.

The entire set is 24 patches, so I'm not going to send them out
individually. I've got the entire set broken out into individual patches
here:

    http://people.redhat.com/rhel6/cifs

This set is nearly identical to the set I tested at Connectathon. There
are just a couple of small differences, but they tested out fine here.

Almost all of these patches are either in Linus' tree already or are in
Steve French's tree. The exception is the last few patches in the
series, which I sent to Steve a few weeks ago but he's stalling on
reviewing them.

RHEL6 probably can't wait any longer (if it's not already too late), so
I'm going ahead and sending them in the expectation that he'll take them
eventually. If he doesn't, I'll post a cleanup patch that fixes it to
match what eventually went upstream.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>