GFS2: Mandatory locking fix
Patchwork-id: 23533
O-Subject: [RHEL 6] GFS2: Mandatory locking fix (bz #571606) - EMBARGOED
Bugzilla: 571606
CVE: CVE-2010-0727
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Benjamin Marzinski <bmarzins@redhat.com>

Hi,

This is the RHEL6 patch for this particular issue. We'll be posting the
fix for various versions of RHEL today and I hope, upstream too shortly.
This has been verified to fix the reproducer in the master bug.

Steve.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>