getting rid of key.h and making it temporary
When building the kernel locally, 'make -rh-key' is done automatically so
kernel.{sec,pub} and crypto/signature/key.h can be created. problem is:
key.h already exists in the repository and git won't ignore changes to it.
Removing it since the stub has no use and the build can fail if the file
doesn't have the key.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>