SELinux: reduce size of access vector hash table
Patchwork-id: 23564
O-Subject: [PATCH RHEL6] SELinux: reduce size of access vector hash table
Bugzilla: 570433
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: James Morris <jmorris@redhat.com>

BZ 570433

We have seen a couple of reports where people are having trouble with
loading SELinux policy most likely due to memory fragmentation.  SELinux
can try to do a kcalloc() when loading policy with size 65536 which almost
always works at boot but can easily fail after a machine has been up for a
while (we've seen it during system install a couple of times).  This patch
cuts the number of hash buckets we use way down and thus cuts the allocation
in down as well.  From 64k allocation to a 16k allocation.  This is not a
really perf critical path so the smaller hash table shouldn't be an issue.
Upstream I plan to be the first user of flex_array.c but for RHEL this seemes
the smallest and most prudent solution.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>