SELinux: reset the security_ops before flushing the avc cache
Patchwork-id: 23565
O-Subject: [PATCH RHEL6] SELinux: reset the security_ops before flushing the avc
	cache
Bugzilla: 572998
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: James Morris <jmorris@redhat.com>

BZ 572998

This is just a TINY possible resource usage waste.  upstream commit
message is as follows

This patch resets the security_ops to the secondary_ops before it flushes
the avc.  It's still possible that a task on another processor could have
already passed the security_ops dereference and be executing an selinux hook
function which would add a new avc entry.  That entry would still not be
freed.  This should however help to reduce the number of needless avcs the
kernel has when selinux is disabled at run time.  There is no wasted
memory if selinux is disabled on the command line or not compiled.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: James Morris <jmorris@namei.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>