selinux/ss: correct size computation
Patchwork-id: 23571
O-Subject: [PATCH RHEL6] security/selinux/ss: correct size computation
Bugzilla: 573000
RH-Acked-by: James Morris <jmorris@redhat.com>

From: Julia Lawall <julia@diku.dk>

BZ 573000

This is not a bug on any of our platforms since sizeof(void *) == sizeof (void **)
but it is right.  The upstream commit message is as follows:

The size argument to kcalloc should be the size of desired structure,
not the pointer to it.

The semantic patch that makes this change is as follows:
(http://coccinelle.lip6.fr/)

// <smpl>
@expression@
expression *x;
@@

x =
 <+...
-sizeof(x)
+sizeof(*x)
...+>// </smpl>

Signed-off-by: Julia Lawall <julia@diku.dk>
Acked-by: Eric Paris <eparis@redhat.com>
Signed-off-by: James Morris <jmorris@namei.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>