ixgbe: update to version 2.0.62-k2
Patchwork-id: 23572
O-Subject: [PATCH RHEL6.0 1/2] ixgbe: update to version 2.0.62-k2
Bugzilla: 462790
RH-Acked-by: John Linville <linville@redhat.com>

All valid upstream fixes through:

	commit 2ea186ae533c7b4f4c56811b69d3e40a6209a9c0
	Author: John Fastabend <john.r.fastabend@intel.com>
	Date:   Sat Feb 27 03:28:24 2010 -0800

	    ixgbe: move TC_PRIO_CONTROL check into ixgbe_select_queue()

I have tested this driver and asked Intel for feedback.  This is close
to what is currently upstream, so I do not expect any issues to be found
that are not present there.

This series will resolve RHBZ 462790.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>