dlm: fix ordering of bast and cast
Patchwork-id: 23663
O-Subject: [RHEL6.0 PATCH 1/3] dlm: fix ordering of bast and cast
Bugzilla: 568102
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

bz 568102

upstream:
http://git.kernel.org/gitweb.cgi?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=7fe2b3190b8b299409f13cf3a6f85c2bd371f8bb

brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=2329735

When both blocking and completion callbacks are queued for lock,
the dlm would always deliver the completion callback (cast) first.
In some cases the blocking callback (bast) is queued before the
cast, though, and should be delivered first.  This patch keeps
track of the order in which they were queued and delivers them
in that order.

This patch also keeps track of the granted mode in the last cast
and eliminates the following bast if the bast mode is compatible
with the preceding cast mode.  This happens when a remotely mastered
lock is demoted, e.g. EX->NL, in which case the local node queues
a cast immediately after sending the demote message.  In this way
a cast can be queued for a mode, e.g. NL, that makes an in-transit
bast extraneous.

Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>