dlm: send reply before bast
Patchwork-id: 23664
O-Subject: [RHEL6.0 PATCH 2/3] dlm: send reply before bast
Bugzilla: 568102
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

bz 568102

upstream:
http://git.kernel.org/gitweb.cgi?p=linux/kernel/git/torvalds/linux-2.6.git;a=com

brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=2329735

When the lock master processes a successful convert or request,
it will process the effects of that change before sending the
reply for the convert/request.  The effects are:

- blocking callbacks (basts) for the newly granted locks
- waiting or converting locks that can now be granted

The cast is queued on the local node when the reply from the lock
master is received.  This means that a lock holder can receive a
bast for a lock mode that is doesn't yet know has been granted.

Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>