dlm: use bastmode in debugfs output
Patchwork-id: 23665
O-Subject: [RHEL6.0 PATCH 3/3] dlm: use bastmode in debugfs output
Bugzilla: 568102
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

bz 568102

upstream:
http://git.kernel.org/gitweb.cgi?p=linux/kernel/git/torvalds/linux-2.6.git;a=com

brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=2329735

The bast mode that appears in the debugfs output should be
useful on both master and process nodes.  lkb_highbast is
currently printed, and is only useful on the master node.
lkb_bastmode is only useful on the process node.  This
patch sets lkb_bastmode on the master node as well, and
uses that value in the debugfs print.

Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>