python: change scripts to use system python instead of env
Patchwork-id: 23706
O-Subject: [PATCH RHEL-6] [scripts] python: change scripts to use system python
	instead of env
Bugzilla: 575965
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=575965

This is on the upstream maintainers queue to push upstream.  He just
hasn't be enable to because he got burned by Linus moving the -rc1 date.

Just a small change to a couple of scripts to go from

 #!/usr/bin/env python

to

 #!/usr/bin/python

This shouldn't effect anyone, unless they don't install python there.

In preparation for python3, Fedora is doing a big push to change the scripts
to use the system python.  This allows developers to put the python3 in
their path without fear of breaking existing scripts.

Now I am pretty sure anyone using python3 for testing purposes will probably
not run any of the scripts I changed, but Fedora has this automated tool
that checks for this stuff so I thought I would try to push it upstream.

Signed-off-by: Don Zickus <dzickus@redhat.com>

--
I don't do enough python to defend this change to much other than it seems
to make sense.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>