zcore: CPU registers are not saved under LPAR
Patchwork-id: 23728
O-Subject: [RHEL6 PATCH 5/7] [s390] zcore: CPU registers are not saved under LPAR
Bugzilla: 575221
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
Dumps created on LPAR with zfcpdump may not contain register information.

To save the registers for all CPUs a sigp "store status" is done that
stores the registers to address absolute zero.  To access storage at
absolute zero, normally the address of the prefix register of the
accessing CPU has to be used. This does not work when large pages are
active (currently only under LPAR).

Note: There is a case where the problem does not occur: When the prefix
      page is allocated at an address under 1MB then the registers are
      saved correctly also without the patch.
      The reason is that the large page mapping starts at 1MB.

Use memcpy_real function instead of memcpy to switch to real mode where
prefixing works.

Bugzilla
--------
BZ 575221
https://bugzilla.redhat.com/show_bug.cgi?id=575221

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.34
Posted to LKML: http://lkml.org/lkml/2010/3/23/72

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>