dasd: fix alignment of transport mode recovery TCW
Patchwork-id: 23730
O-Subject: [RHEL6 PATCH 7/7] [s390] dasd: fix alignment of transport mode
	recovery TCW
Bugzilla: 575824
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
A recovery TCW fails with CS: 0x20 and schxs: 0x06

When the DASD device driver uses the transport mode (High Performance
FICON) and there is a request that needs error recovery, then we create
a recovery requests with a fresh recovery TCW.  Every TCW needs to be
aligned on a 64 byte boundary, but for the recovery TCWs we fail to do
the proper alignment.  The first recovery request for a device has
proper alignment by coincident, so most of the times recovery will work.
But if there is more then one recovery request at the same time, then the
recovery will fail and result in an I/O error.

This patch provides proper alignment for the TCW.

Bugzilla
--------
BZ 575824
https://bugzilla.redhat.com/show_bug.cgi?id=575824

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.34
Posted to LKML: http://lkml.org/lkml/2010/3/23/72

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>