cciss: add 30 second initial timeout wait on controller reset
Patchwork-id: 23731
O-Subject: [RHEL6 PATCH] cciss: add 30 second initial timeout wait on controller
	reset
Bugzilla: 574094
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz#574094

This is a backport of http://ns.spinics.net/lists/linux-scsi/msg42595.html
Upstream status - only posted on linux-scsi

I've tested this on i686 with the '5i' controller together
with the latest kexec-tools. With all this kdump works again.
Explanation why is this needed from the patch:
/* The HP Smart Array 5i Controller needs
 * at least 20 seconds before first status checking
 * set it to 30 seconds for this controller to be sure */
(The timeout is added only during kdump)

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>