Intel ICH9 workaround for HPET timer issue on IbexPeak Platform
Patchwork-id: 23740
O-Subject: [RHEL 6.0 PATCH] bz502629 Intel ICH9 workaround for HPET timer issue
	on IbexPeak Platform
Bugzilla: 502629
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Description of problem:

HPET timer issue on Calpella Platform
There is an issue found on HPET timers on Ibexpeak based platforms (NHM Desktop
and Mobile platforms). The problem is that Read of HPET comparator immediately
after a Write, returns the old value instead of recently written value.
Experiments showed that the problem was only with Comparator 0 in HPET and a
second read following the write gives the right value. We have a sighting open
with the hardware team, and hardware folks have acknowledged that there is a
"small window" in which a read after a write can return old value. We will
likely have this as an errata for this chipset and current workaround for it is
to read HPET twice.

Spec update:

http://www.intel.com/Assets/PDF/specupdate/319974.pdf

And public discussion of the problem and workaround
http://marc.info/?l=linux-kernel&m=126712404431804&w=2  <http://marc.info/?l=linux-kernel&m=126712404431804&w=2>

Upstream status:
Patch public available on LKML, don't know commit id yet.

Testing status:
The patch has been tested by Intel and fixed the problem.
I'm waiting on the brew and will give it a run later.

Brew info:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2338205

Please review and ACK.

Thanks,
Luming

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>