ahci: support FIS-based switching
Patchwork-id: 23743
O-Subject: [RHEL6 PATCH] ahci: support FIS-based switching
Bugzilla: 463152
RH-Acked-by: Jeff Garzik <jgarzik@redhat.com>

Some AHCI controllers support FIS-based switching which allows the
host to issue commands to a device while there are still outstanding
commands to other devices connected to the same host port through
a port multiplier. NCQ and non-NCQ commands cannot be in-flight to
the same device, but they can be in-flight to different devices
connected to the port multiplier. This patch is provided by Shane
Huang, and has been tested by AMD.

Upstream commits:

commit d6ef31539d715db7a0eb2d67a1a008c9d8ccf059
Author: Shane Huang <shane.huang@amd.com>
Date:   Wed Dec 9 17:23:04 2009 +0800

    ahci: Implement SATA AHCI FIS-based switching support

commit 5db5b0215af94a36d4bf10900ff9707b6d5c1610
Author: Shane Huang <shane.huang@amd.com>
Date:   Tue Mar 16 18:08:55 2010 +0800

    ahci: pp->active_link is not reliable when FBS is enabled

    pp->active_link is not reliable when FBS is enabled.
    Both PORT_SCR_ACT and PORT_CMD_ISSUE should be checked
    because mixed NCQ and non-NCQ commands may be in flight.

This is for BZ 463152, please review and ACK.

Thanks,
David

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>