Add padding to mm structures allow future patches during the RHEL6 life
Bugzilla: 554511

We're working with a partner who is developing performance enhancements
for a variety of layered products.  These enhancements include both user
and kernel code which significantly improve the performance of large
multi-threaded( especially java) applications.  This partner is working
with the upstream community on developing the kernel enhancements which
are generic enough to useful to everyone.  They expect the kernel part
to undergo several iterations with the upstream community during the
development phase.  Since we would like the opportunity to include these
in RHEL6 during the 7-year life cycle without breaking the kABI, I am
proposing adding padding a couple kernel data structures.   The attached
patch adds 2 "void*"s to the vm_area_struct and one "void*" to the
mm_struct.

Addresses future expansion possibility for BZ554511

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>