e100: fix the 'size' argument passed to pci_pool_create()
Patchwork-id: 23790
O-Subject: [RHEL6.0 PATCH] e100: fix the 'size' argument passed to
	pci_pool_create()
Bugzilla: 576887
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Resolves RHBZ 576887.

If the size of the TX ring is increased for the e100 driver, the system is
liable to panic. This is because the size passed to pci_pool_create() reflected
the current default size and not the max possible size.

Backport of upstream patch:
http://git.kernel.org/linus/211a0d941b1924e667483f822a55e2cc694cd212

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2343486

Tested by me on pogolinux-2.rhts.eng.rdu.redhat.com.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>