update fibre channel layer
Patchwork-id: 23792
O-Subject: [PATCH RHEL 6]: update fibre channel layer
Bugzilla: 571824
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for BZ 571824.

The bz was origingally only going to bring in a fix for the
port_sate race that was being hit, but while debugging it and
checking that code was upstream I thought it was be best to bring in
some other fixes. This patch syncs the FC class with what is in the
scsi maintainer's tree.

There is a fix for the race we were hitting where when
the host lock was dropped the remote port state changed
and instead of fixing the locking or rechecking the state
we just did a BUG_ON.
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=8798a694da59486e4a3ff0abeec183202fb34c20

It includes a fix for a vport deletion race:
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=0d9dc7c8b9b7fa0f53647423b41056ee1beed735

And it includes a fix for how we manage the timers used
by multipath.
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=f2818663c82b7297ff4aa38cbddb870dc02f7104

I have tested the first and last patch by disabling/enabling ports
with different timeouts. I have not been able to recreate the second
problem because I am having trouble setting up rports on our switch.
I am sending it because it is upstream and after reviewing the code
it looks correct. I am also hoping since it is going in now partners
will be testing.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>