zfcp: Remove lock dependency on CCW remove
Patchwork-id: 23804
O-Subject: [RHEL6 PATCH 1/3] [s390] zfcp: Remove lock dependency on CCW remove
Bugzilla: 576860
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
During the removal of a CCW device the config_mutex has to be
held to maintain the configuration consistency.  However, a few
process steps do fetch other locks which might interfere with
the already held mutex.

The double protection is not required, so remove it.

Bugzilla
--------
BZ 576860
https://bugzilla.redhat.com/show_bug.cgi?id=576860

Upstream status of the patch
----------------------------
The patch is specific to the 2.6.32 kernel and is not
applicable to the upstream kernel.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>