exec: refactor how usermodehelpers work and modify core_pipe recursion check
Patchwork-id: 23840
O-Subject: [RHEL6 PATCH] exec: refactor how usermodehelpers work and modify
	core_pipe recursion check
Bugzilla: 557387
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Dave Anderson <anderson@redhat.com>

Hey all-
	This is a bit of a long story, but the short version is that about 8
months ago I changed how the coredump-to-a-pipe recursion check worked upstream,
which mandated a change in how utilities like abrt worked.  the Abrt guys
apparently only discovered this two months ago and asked it to be changed back.
we couldn't do that without re-introducing the bug that I had initially fixed,
so after some back and forth we came up with this solution, which Oleg N. and I
also used as an opportunity to clean up how the usermodehelper code was
implemented.  Its in mm and partially in -stable now, and we need it so Abrt can
work with RHEL6 without modification.  Tested by the Abrt crew, myself, and a
few people at canonical, works well.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>